fix: run agentDir migration on TUI startup too (fixes /resume)
The migration was only running in the daemon path. When using the TUI
directly (not via runner), /resume couldn't find sessions because they
were still at ~/.pizzapi/agent/sessions/ and the migration hadn't run.

Extract migrateAgentDir() into shared migrations.ts module and call it
from both daemon.ts and index.ts (TUI interactive path).
